*** empty log message ***

+++ gauss_solve.py
@@ -15,7 +15,7 @@ n = len(A)
 flat_array_2d = [item for row in A for item in row]
 
 # Convert to a ctypes array
-c_array_2d = (ctypes.c_int * len(flat_array_2d))(*flat_array_2d)
+c_array_2d = (ctypes.c_double * len(flat_array_2d))(*flat_array_2d)
 
 # Define the function signature
 lib.lu_in_place.argtypes = (ctypes.c_int, ctypes.POINTER(ctypes.POINTER(ctypes.c_double)))
